帮助位置从右上角移动到公告栏
